[FE] REFACT: AGUReturn 말고 AGUPage에서 대여정보 요청 #1737
